Adds model admission webhook
Implements a validating and mutating admission webhook for Model resources.

This webhook adds defaulting logic, such as setting
`scaleToZeroGracePeriod` and `costExpansionRatePercent` if autoscaling is enabled,
and performs validations, like ensuring that `minReplicas` is less than
or equal to `maxReplicas` and worker types are correctly set based on backend.

Also adds a Dockerfile and manifests for deploying the webhook.