Support Helm (#165)
* init helm

* fix according to AI

* fix bug

* fix bug

* fix bug

* fix bug

* change app.kubernetes.io/instance to app.kubernetes.io/component

* Transfer to subcharts

* Add helm check CI

* Keep model controller configmap namespace same with helm namespace

* make lint

* Change way to install helm

* Add README.md

* Add secrets for downloader

* Add infer gateway

* Add condition for subchart

* Fix according to AI

* Fix according to AI

* Fix conflict

* Add document for how to only install specific subchart

* Add document for how to only install specific subchart

* Print error in log

* update readme.md for helm

* udpate readme.md for helm

* udpate readme.md for helm

* udpate readme.md for helm

* udpate readme.md for helm

* Rename chart files and update chart name to matrixinfer

* Rename chart files and update chart name to matrixinfer

* update readme.md

* add Certificate and Issuer template for the usage of cert manager

* parameterize cert manager

* make webhook conditional

* Renames `registryWebhook` to `webhook`

Updates values.yaml parameters from `registryWebhook` to `webhook` for clarity and consistency.

This change simplifies the configuration and improves the readability of the Helm chart.

* Update CRD output paths and rename related files for matrixinfer chart

* Update README.md

* Refactors webhook Dockerfile for build context

Updates the Dockerfile to copy specific source code directories and files instead of the entire context. This improves build times and reduces the size of the resulting image by excluding unnecessary files.

Adds a .dockerignore file to further refine the build context and exclude common development-related files and directories, further optimizing the build process and reducing the image size.

* Renames gateway subchart to networking

Updates the gateway subchart's name to networking to better reflect its functionality.

Adjusts references in the parent chart and documentation to reflect the rename.

* fix naming

* render fix

* Fix bug

---------

Co-authored-by: Lu Ma <malu17@huawei.com>