[Webhook][Certificates] Certs and CA of webhook can now be set up manually instead of using cert manager (#214)
* Enables fallback TLS configuration for webhooks

Allows manual configuration of TLS certificates for webhooks
when cert-manager is not enabled. This provides a fallback
mechanism for environments where cert-manager is not available.

Adds documentation explaining the different configuration options.

* add temporary file for test

* update name

* Simplifies webhook names and clarifies install instructions

Renames mutating and validating webhook configurations for clarity and consistency.

Updates the README to clarify helm installation instructions, specifically adding the `-n ` flag to the installation and uninstallation commands. This ensures users understand the importance of specifying a namespace during deployment.

* Removes manual certificate setup guide.

Deletes the guide for manual certificate setup. The
guide is no longer needed.

* udpate README.md

* Ensures caBundle is required when cert-manager is disabled

Makes the `caBundle` value required for the registry and workload webhooks when cert-manager is disabled.
This prevents webhook creation from failing silently when a `caBundle` is not provided when `cert-manager` is disabled.

* Refactors webhook caBundle configuration

Moves the caBundle configuration for webhooks to a global scope.

This change centralizes the caBundle setting, making it easier to manage and ensuring consistency across different charts. It updates the required annotation for caBundle.

* Removes redundant readme content

* Revert "Removes redundant readme content"

This reverts commit 2192f949071677ad3eab9af6ede392e12b653847.

* Documents webhook certificate configuration

Adds documentation detailing how to configure TLS certificates for webhooks, including using cert-manager and manual configuration options with CA bundle setup.

Removes duplicated documentation from the root README.md and centralizes it in the chart's README.md.

* Adds registry webhook documentation

Adds a comprehensive README document for the registry webhook.

This document provides information on validation rules, default values applied by the mutator webhook, webhook configuration details, and instructions for extending the webhooks with new validation rules, default values, and resource support.

* Updates struct field alignment

Addresses a minor formatting issue by adjusting the field alignment within the `modelInferConfig` struct for improved code readability.