[Docs] Adds test workflow for docs & Write docs for cert manager integration chapter (#258)
* Adds documentation tests workflow

Sets up a new GitHub Actions workflow to run tests specifically for the Docusaurus documentation located in the `docs/matrixinfer` directory.

This workflow triggers on pull requests and pushes to the `main` branch, but only when changes are made within the `docs/matrixinfer` directory. It installs dependencies, performs type checking, and builds the documentation to ensure its integrity.

* Adds cert-manager integration documentation

Adds documentation for integrating with cert-manager to automate SSL/TLS certificate management for MatrixInfer components, including configuration steps, resource details, best practices, and troubleshooting tips.

Also, modifies the gateway issuer and certificate names in the Helm chart to use the chart name instead of the fullname.

* Removes best practices section.

Removes the "Best Practices" section from the cert-manager documentation.
The content was deemed less critical in the current scope of the document.