Generates CLI documentation using cobra-doc
Moves CLI documentation generation from a cobra command to a standalone tool.
This allows generating the documentation as part of the build process
and ensures up-to-date CLI documentation in the repository.

The new tool uses `cobra-doc` to generate Markdown documentation for the CLI,
which is then placed in the `docs/matrixinfer/docs/reference/cli` directory.
A Makefile entry automates this process.

A prettier config is also added to ensure consistent formatting.

The old `doc` command is removed from the CLI.