[Docs] miscellaneous small updates (#340)
* switch to lightbox image for model infer svg

* Update CLI documentation

* Updates quick start guide and link color

Refactors the quick start guide to improve clarity and correct the kubectl command.

Updates the default link color to the primary color for better visibility.

* Adds line numbers to YAML code block

Enables line numbers for the YAML code block in the quick start guide.

This improves readability and makes it easier for users to reference specific lines when working with the example configuration.

* update dark mode image color

* Removes obsolete container descriptions

Removes descriptions for Init and Sidecar containers from the architecture documentation.

These container types are no longer relevant to the current MatrixInfer architecture.